# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PizzaHub API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/addresses/{address}": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Получить адрес по ID",
                "operationId": "cef120083854225004b7b0ca60846aa3",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 3
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Мира 25"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Минск"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Address retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Адрес не найден"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Обновить адрес",
                "description": "Обновление адреса по его ID",
                "operationId": "updateAddress",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID адреса для обновления",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 23
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address_line_1",
                                    "city"
                                ],
                                "properties": {
                                    "address_line_1": {
                                        "description": "Адресная строка 1",
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Мира 25"
                                    },
                                    "city": {
                                        "description": "Город",
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "Минск"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Адрес обновлен успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 23
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Мира 25"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Минск"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Addresses retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Адрес с таким ID не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Address with ID 21 not found"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "Ошибка 1",
                                                "Ошибка 2"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Удалить адрес",
                "description": "Удаление адреса по его ID",
                "operationId": "deleteAddress",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "description": "ID адреса для удаления",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Адрес успешно удален",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 24
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Советская 1"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Москва"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Address with ID 2 deleted successfully."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Адрес с таким ID не найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Address with ID 21 not found"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 404
                                        },
                                        "details": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "example": [
                                                "Details about the error"
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/addresses": {
            "get": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Получить список адресов",
                "operationId": "c2cf707eeb35a73be1c94dc850806f43",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "address": {
                                                        "type": "string",
                                                        "example": "Советская 1"
                                                    },
                                                    "city": {
                                                        "type": "string",
                                                        "example": "Москва"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Addresses retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Addresses"
                ],
                "summary": "Создать новый адрес",
                "operationId": "701875ed84c11c469a175891494292fc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "address_line_1",
                                    "city"
                                ],
                                "properties": {
                                    "address_line_1": {
                                        "type": "string",
                                        "example": "Советская 1"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Москва"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Адрес создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 24
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Советская 1"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Москва"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "New address created."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 422
                                        },
                                        "details": {
                                            "properties": {
                                                "address_line_1": {
                                                    "type": "string",
                                                    "example": "The address line 1 field is required."
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}